services/search-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic[email]==2.5.3
bcrypt==4.1.2
passlib==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
email-validator==2.1.0
